[WSCH-275] A2 RTP Release initial merge
[WSCH-275] package.yml file update for Misra build

[WSCH-275] MISRA Check Stage issue fix-1

[WSCH-275] MISRA Check stage 2 fix 1

[WSCH-275] Ble_sensor_file addition

[WSCH-275] MISRAC Check Stage 2 fix 3

[WSCH-275] path loss & ANCS application added

[WSCH-275] MISRAC Check stage 2 fix 4

WSCH-275 MISRAC Check stage 2 fix 5

[WSCH-275] MISRAC Check Stage 2 fix 5

[WSCH-275] MISRA C Check Stage 3 fix 1

[WSCH-275] MISRA C Check stage 2 fix 6

[WSCH-275] Pathloss application and Misra check stage 3 fix 1

[WSCH-275] MISRAC Check stage 3 fix 3

[WSCH-275] MISRAC Check stage 3 fix 4

[WSCH-275] MISRAC check on dev branch

[WSCH-275] WBZ451 A2 RTP Release

+++ apps/ble/building_blocks/central/central_conn/firmware/src/config/default/osal/osal_freertos_extend.h
@@ -16,7 +16,7 @@
 
 // DOM-IGNORE-BEGIN
 /*******************************************************************************
-* Copyright (C) 2018 Microchip Technology Inc. and its subsidiaries.
+* Copyright (C) 2022 Microchip Technology Inc. and its subsidiaries.
 *
 * Subject to your compliance with these terms, you may use Microchip software
 * and any derivatives exclusively with Microchip products. It is your
@@ -155,7 +155,7 @@ OSAL_RESULT OSAL_QUEUE_AddToSet(OSAL_QUEUE_SET_MEMBER_HANDLE_TYPE *queSetMember,
 *@retval OSAL_RESULT_TRUE    - A queue had been created
 *        OSAL_RESULT_FALSE   - Queue creation failed
 */
-OSAL_RESULT OSAL_QUEUE_SelectFromSet(OSAL_QUEUE_SET_MEMBER_HANDLE_TYPE *queSetMember, OSAL_QUEUE_SET_HANDLE_TYPE *queSetID, uint16_t waitMS);
+OSAL_RESULT OSAL_QUEUE_SelectFromSet(OSAL_QUEUE_SET_MEMBER_HANDLE_TYPE *queSetMember, OSAL_QUEUE_SET_HANDLE_TYPE *queSetID, uint32_t waitMS);
 
 // *****************************************************************************
 /**
@@ -177,7 +177,7 @@ OSAL_RESULT OSAL_QUEUE_SelectFromSet(OSAL_QUEUE_SET_MEMBER_HANDLE_TYPE *queSetMe
 *@retval OSAL_RESULT_TRUE    - Item copied to the queue
 *        OSAL_RESULT_FALSE   - Item not copied to the queue or timeout occurred
 */
-OSAL_RESULT OSAL_QUEUE_Send(OSAL_QUEUE_HANDLE_TYPE *queID, void *itemToQueue, uint16_t waitMS);
+OSAL_RESULT OSAL_QUEUE_Send(OSAL_QUEUE_HANDLE_TYPE *queID, void *itemToQueue, uint32_t waitMS);
 
 // *****************************************************************************
 /**
@@ -224,7 +224,7 @@ OSAL_RESULT OSAL_QUEUE_SendISR(OSAL_QUEUE_HANDLE_TYPE *queID, void  *itemToQueue
 *        OSAL_RESULT_FALSE   - An item was not successfully received from the queue
 *                          or timeout occurred
 */
-OSAL_RESULT OSAL_QUEUE_Receive(OSAL_QUEUE_HANDLE_TYPE *queID, void  *pBuffer, uint16_t waitMS);
+OSAL_RESULT OSAL_QUEUE_Receive(OSAL_QUEUE_HANDLE_TYPE *queID, void  *pBuffer, uint32_t waitMS);
 
 // *****************************************************************************
 /**
